+++ lib/src/mysql_protocol/packet/packet_handshake_response_41.dart
@@ -4,6 +4,13 @@ import 'package:buffer/buffer.dart';
 import 'package:mysql_client/mysql_protocol.dart';
 import 'package:mysql_client/mysql_protocol_extension.dart';
 
+const _supportedCapabitilies = mysqlCapFlagClientProtocol41 |
+    mysqlCapFlagClientSecureConnection |
+    mysqlCapFlagClientPluginAuth |
+    mysqlCapFlagClientPluginAuthLenEncClientData |
+    mysqlCapFlagClientMultiStatements |
+    mysqlCapFlagClientMultiResults;
+
 class MySQLPacketHandshakeResponse41 extends MySQLPacketPayload {
   int capabilityFlags;
   int maxPacketSize;
@@ -44,10 +51,7 @@ class MySQLPacketHandshakeResponse41 extends MySQLPacketPayload {
     );
 
     return MySQLPacketHandshakeResponse41(
-      capabilityFlags: mysqlCapFlagClientProtocol41 |
-          mysqlCapFlagClientSecureConnection |
-          mysqlCapFlagClientPluginAuth |
-          mysqlCapFlagClientPluginAuthLenEncClientData,
+      capabilityFlags: _supportedCapabitilies,
       maxPacketSize: 50 * 1024 * 1024,
       authPluginName: initialHandshakePayload.authPluginName!,
       characterSet: initialHandshakePayload.charset,
@@ -74,10 +78,7 @@ class MySQLPacketHandshakeResponse41 extends MySQLPacketPayload {
     );
 
     return MySQLPacketHandshakeResponse41(
-      capabilityFlags: mysqlCapFlagClientProtocol41 |
-          mysqlCapFlagClientSecureConnection |
-          mysqlCapFlagClientPluginAuth |
-          mysqlCapFlagClientPluginAuthLenEncClientData,
+      capabilityFlags: _supportedCapabitilies,
       maxPacketSize: 50 * 1024 * 1024,
       authPluginName: initialHandshakePayload.authPluginName!,
       characterSet: initialHandshakePayload.charset,