+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.22
+
+- Check server supports SSL
+- Add support for multiple statements
+
 ## 0.0.21
 
 - Fix _lastError reset in _forceClose() and used after